Chameleon {seriation}	R Documentation

2D Data Sets used for the CHAMELEON Clustering Algorithm

Description

Several 2D data sets used to evaluate the CHAMELEON clustering alogrithm in the paper by Karypis et al (1999) and used by iVAT, an ordering-based tool to asses cluster tendency (Havens and Bezdek, 2012).

Usage

data(Chameleon)

Format

chameleon_ds4: The format is a 8,000 x 2 data.frame.

chameleon_ds5: The format is a 8,000 x 2 data.frame.

chameleon_ds7: The format is a 10,000 x 2 data.frame.

chameleon_ds8: The format is a 8,000 x 2 data.frame.

Source

The data was obtained from http://glaros.dtc.umn.edu/gkhome/cluto/cluto/download

References

Karypis, G., EH. Han, V. Kumar (1999): CHAMELEON: A Hierarchical Clustering Algorithm Using Dynamic Modeling, IEEE Computer, 32(8): 68–75.

Havens, T.C. and Bezdek, J.C. (2012): An Efficient Formulation of the Improved Visual Assessment of Cluster Tendency (iVAT) Algorithm, IEEE Transactions on Knowledge and Data Engineering, 24(5), 813–822.

Examples

data(Chameleon)

plot(chameleon_ds4, cex=.1)

[Package seriation version 1.2-3 Index]