Registry for Criterion Methods

A registry to manage methods used by criterion() to calculate a criterion value given data and a permutation.

Usage

registry_criterion

list_criterion_methods(kind)

get_criterion_method(kind, name)

set_criterion_method(kind, name, fun, description = NULL, merit = NA, ...)

# S3 method for criterion_method
print(x, ...)

Format

An object of class criterion_registry (inherits from registry) of length 20.

Arguments

kind

the data type the method works on. For example, "dist", "matrix" or "array".

name

the name for the method used to refer to the method in the function criterion().

fun

a function containing the method's code.

description

a description of the method. For example, a long name.

merit

a boolean indicating if the criterion measure is a merit (TRUE) or a loss (FALSE) measure.

...

further information that is stored for the method in the registry.

x

an object of class "criterion_method" to be printed.

Value

list_criterion_method() results is a vector of character strings with the names of the methods used for criterion().

get_criterion_method() returns a given method in form of an object of class "criterion_method".

Details

All methods below are convenience methods for the registry named registry_criterion.

list_criterion_method() lists all available methods for a given data type (kind). The result is a vector of character strings with the short names of the methods. If kind is missing, then a list of methods is returned.

get_criterion_method() returns information (including the implementing function) about a given method in form of an object of class "criterion_method".

With set_criterion_method() new criterion methods can be added by the user. The implementing function (fun) needs to have the formal arguments x, order, ..., where x is the data object, order is an object of class ser_permutation_vector and ... can contain additional information for the method passed on from criterion(). The implementation has to return the criterion value as a scalar.

See also

Author

Michael Hahsler

Examples

## the registry
registry_criterion
#> An object of class 'registry' with 20 entries.

# List all criterion calculation methods by type
list_criterion_methods()
#> $dist
#>  [1] "2SUM"              "AR_deviations"     "AR_events"        
#>  [4] "BAR"               "Cor_R"             "Gradient_raw"     
#>  [7] "Gradient_weighted" "Inertia"           "LS"               
#> [10] "Lazy_path_length"  "Least_squares"     "ME"               
#> [13] "Moore_stress"      "Neumann_stress"    "Path_length"      
#> [16] "RGAR"             
#> 
#> $matrix
#> [1] "Cor_R"          "ME"             "Moore_stress"   "Neumann_stress"
#> 

# List methods for matrix
list_criterion_methods("matrix")
#> [1] "Cor_R"          "ME"             "Moore_stress"   "Neumann_stress"

get_criterion_method("dist", "AR_d")
#> name:        AR_deviations
#> kind:        dist
#> description: Anti-Robinson deviations
#> merit:       FALSE

# Define a new method (sum of the diagonal elements)

## 1. implement a function to calculate the measure
criterion_method_matrix_foo <- function(x, order, ...) {
if(!is.null(order)) x <- permute(x,order)
    sum(diag(x))
}

## 2. Register new method
set_criterion_method("matrix", "DiagSum", criterion_method_matrix_foo,
    description = "Calculated the sum of all diagonal entries", merit = FALSE)

list_criterion_methods("matrix")
#> [1] "Cor_R"          "DiagSum"        "ME"             "Moore_stress"  
#> [5] "Neumann_stress"
get_criterion_method("matrix", "DiagSum")
#> name:        DiagSum
#> kind:        matrix
#> description: Calculated the sum of all diagonal entries
#> merit:       FALSE

## 3. use all criterion methods (including the new one)
criterion(matrix(1:9, ncol = 3))
#>          Cor_R        DiagSum             ME   Moore_stress Neumann_stress 
#>    -0.09301487    15.00000000   340.00000000   280.00000000   120.00000000