Registry for Seriation Methods

A registry to manage methods used by seriate().

Usage

registry_seriate

list_seriation_methods(kind)

get_seriation_method(kind, name)

set_seriation_method(
  kind,
  name,
  definition,
  description = NULL,
  control = list(),
  ...
)

# S3 method for seriation_method
print(x, ...)

Format

An object of class seriation_registry (inherits from registry) of length 48.

Arguments

kind

the data type the method works on. For example, "dist", "matrix" or "array". If missing, then methods for any type are shown.

name

the name for the method used to refer to the method in seriate().

definition

a function containing the method's code.

description

a description of the method. For example, a long name.

control

a list with control arguments and default values.

...

further information that is stored for the method in the registry.

x

an object of class "seriation_method" to be printed.

Value

list_seriation_method() result is a vector of character strings with the names of the methods. These names are used for methods in seriate().

get_seriation_method() returns a given method in form of an object of class "seriation_method".

Details

The functions below are convenience function for the registry registry_seriate.

list_seriation_method() lists all available methods for a given data type (kind) (e.g., "dist", "matrix"). The result is a vector of character strings with the method names that can be used in function seriate(). If kind is missing, then a list of methods is returned.

get_seriation_method() returns detailed information for a given method in form of an object of class "seriation_method". The information includes a description, parameters and the implementing function.

With set_seriation_method() new seriation methods can be added by the user. The implementing function (definition) needs to have the formal arguments x, control, where x is the data object and control contains a list with additional information for the method passed on from seriate(). The implementation has to return a list of objects which can be coerced into ser_permutation_vector objects (e.g., integer vectors). The elements in the list have to be in corresponding order to the dimensions of x.

See also

Author

Michael Hahsler

Examples

# Registry
registry_seriate
#> An object of class 'registry' with 48 entries.

# List all seriation methods by type
list_seriation_methods()
#> $array
#> [1] "Identity" "Random"   "Reverse" 
#> 
#> $dist
#>  [1] "ARSA"          "BBURCG"        "BBWRCG"        "GW"           
#>  [5] "GW_average"    "GW_complete"   "GW_single"     "GW_ward"      
#>  [9] "HC"            "HC_average"    "HC_complete"   "HC_single"    
#> [13] "HC_ward"       "Identity"      "MDS"           "MDS_angle"    
#> [17] "MDS_metric"    "MDS_nonmetric" "OLO"           "OLO_average"  
#> [21] "OLO_complete"  "OLO_single"    "OLO_ward"      "QAP_2SUM"     
#> [25] "QAP_BAR"       "QAP_Inertia"   "QAP_LS"        "R2E"          
#> [29] "Random"        "SA"            "SPIN_NH"       "SPIN_STS"     
#> [33] "Spectral"      "Spectral_norm" "TSP"           "VAT"          
#> 
#> $matrix
#> [1] "BEA"       "BEA_TSP"   "CA"        "Heatmap"   "Identity"  "PCA"      
#> [7] "PCA_angle" "Random"    "Reverse"  
#> 

# List methods for matrix seriation
list_seriation_methods("matrix")
#> [1] "BEA"       "BEA_TSP"   "CA"        "Heatmap"   "Identity"  "PCA"      
#> [7] "PCA_angle" "Random"    "Reverse"  

get_seriation_method(name = "BEA")
#> name:        BEA
#> kind:        matrix
#> description: Bond Energy Algorithm (BEA; McCormick 1972) to maximize the Measure of Effectiveness of a non-negative matrix.
#> control (default values):
#>   istart jstart rep
#> 1      0      0   1

# Example for defining a new seriation method (reverse identity function for matrix)

# 1. Create the seriation method
seriation_method_reverse <- function(x, control) {
   # return a list of order vectors, one for each dimension
   list(seq(nrow(x), 1), seq(ncol(x), 1))
}

# 2. Register new method
set_seriation_method("matrix", "Reverse", seriation_method_reverse,
    description = "Reverse identity order", control = list())

list_seriation_methods("matrix")
#> [1] "BEA"       "BEA_TSP"   "CA"        "Heatmap"   "Identity"  "PCA"      
#> [7] "PCA_angle" "Random"    "Reverse"  
get_seriation_method("matrix", "reverse")
#> name:        Reverse
#> kind:        matrix
#> description: Reverse identity order
#> control: no parameters registered.

# 3. Use the new seriation methods
seriate(matrix(1:12, ncol=3), "reverse")
#> object of class ‘ser_permutation’, ‘list’
#> contains permutation vectors for 2-mode data
#> 
#>   vector length seriation method
#> 1             4          Reverse
#> 2             3          Reverse