seriation: Infrastructure for Ordering Objects Using Seriation

Infrastructure for ordering objects with an implementation of several seriation/sequencing/ordination techniques to reorder matrices, dissimilarity matrices, and dendrograms. Also provides (optimally) reordered heatmaps, color images and clustering visualizations like dissimilarity plots, and visual assessment of cluster tendency plots (VAT and iVAT). Hahsler et al (2008) doi:10.18637/jss.v025.i03 .

Key functions

Seriation: seriate(), criterion(), get_order(), permute()

Visualization: pimage(), bertinplot(), hmap(), dissplot(), VAT()

Available seriation methods and criteria

A list with the implemented seriation methods

A visual comparison between seriation methods

A list with the implemented seriation criteria

Quickstart guides

How to reorder heatmaps

How to reorder correlation matrices

How to evaluate clusters using dissimilarity plots

References

Michael Hahsler, Kurt Hornik, and Christian Buchta. Getting things in order: An introduction to the R package seriation. Journal of Statistical Software, 25(3):1–34, March 2008. doi:10.18637/jss.v025.i03

See also

Author

Maintainer: Michael Hahsler mhahsler@lyle.smu.edu (ORCID) [copyright holder]

Authors:

Michael Hahsler mhahsler@lyle.smu.edu (ORCID) [copyright holder]

Christian Buchta [copyright holder]

Kurt Hornik (ORCID) [copyright holder]

Other contributors:

David Barnett [contributor]

Michael Brusco [contributor, copyright holder]

Michael Friendly [contributor]

Hans-Friedrich Koehn [contributor, copyright holder]

Fionn Murtagh [contributor, copyright holder]

Stephanie Stahl [contributor, copyright holder]