Zoo {seriation}	R Documentation

Zoo Data Set

Description

A database containing characteristics of different animals. The database was created and donated by Richard S. Forsyth and is available from the UCI Machine Learning Repository (Newman et al, 1998).

Usage

data("Zoo")

Format

A data frame with 101 observations on the following 17 variables.

hair

{0, 1}

feathers

{0, 1}

eggs

{0, 1}

milk

{0, 1}

airborne

{0, 1}

aquatic

{0, 1}

predator

{0, 1}

toothed

{0, 1}

backbone

{0, 1}

breathes

{0, 1}

venomous

{0, 1}

fins

{0, 1}

legs

Numeric (set of values: {0, 2, 4, 5, 6, 8})

tail

{0, 1}

domestic

{0, 1}

catsize

{0, 1}

class

a factor with levels amphibian bird fish insect invertebrate mammal reptile

Source

D.J. Newman, S. Hettich, C.L. Blake and C.J. Merz (1998): UCI Repository of machine learning databases, http://www.ics.uci.edu/~mlearn/MLRepository.html, University of California, Irvine, Dept. of Information and Computer Sciences.

Examples

data("Zoo")
x <- scale(Zoo[, -17])

d <- dist(x)
pimage(d)

order <- seriate(d, method = "tsp")
pimage(d, order)

[Package seriation version 1.2-3 Index]