register_DendSer {seriation}	R Documentation

Register Seriation Methods from Package DendSer

Description

Register the DendSer dendrogram seriation method and the ARc criterion (Earle and Hurley, 2015).

Usage

register_DendSer()

Details

Registers the method "DendSer" for seriate. DendSer is a fast heuristic for reordering dendrograms developed by Earle and Hurley (2015) able to use different criteria. control for seriate with method "DendSer" accepts the following parameters:

"h" or "method"

A dendrogram or a method for hierarchical clustering (see hclust). Default: complete-link.

"criterion"

A seriation criterion to optimize (see list_criterion_methods("dist")). Default: "BAR" (Banded anti-Robinson from with 20% band width).

"verbose"

print progress information.

"DendSer_args"

additional arguments for DendSer.

For convenience the following methods (for differnt cost functions) are also provided: "DendSer_ARc" (anti-robinson form), "DendSer_BAR" (banded anti-Robinson form), "DendSer_LS" (leaf seriation), "DendSer_PL" (path length).

Note: Package DendSer needs to be installed.

Author(s)

Michael Hahsler based on code by Catherine B. Hurley and Denise Earle

References

D. Earle, C. B. Hurley (2015): Advances in dendrogram seriation for application to visualization. Journal of Computational and Graphical Statistics, 24(1), 1–25.

See Also

seriate, DendSer in DendSer.

Examples

## Not run: 
register_DendSer()
list_seriation_methods("dist")

d <- dist(random.robinson(20, pre=TRUE))

## use Banded AR form with default clustering (complete-link)
o <- seriate(d, "DendSer_BAR")
pimage(d, o)

## use different hclust method (Ward) and AR as the cost function for
## dendrogram reordering
o <- seriate(d, "DendSer", control = list(method = "ward.D2", criterion = "AR"))
pimage(d, o)

## End(Not run)

[Package seriation version 1.2-3 Index]