uniscale {seriation}	R Documentation

Unidimensional Scaling from Seriation Results

Description

Performs (approximate) unidimensional scaling by first performin seriation to obtain a permutation and the using the permutation to calulate the configuration.

Usage

uniscale(d, order = NULL, method = "QAP_LS", rep = 10, ...)

Arguments

d
a dissimilarity matrix.

order
a precomputed permutation (configuration) order. If NULL, then seriation is performed using the method specified in method.

method
seriation method used if o is NULL.

rep
Number of repetitions of the seriation heuristic.

...
additional arguments are passed on to the seriation method.

Details

Uses the method describes in Maier and De Leeuw (2015) to calculate the minimum stress configuration for either a given configuration/permutation/order or for a permutation computed via a seriation method.

The code is similar to uniscale in smacof, but scales to larger datasets since it does not check all permutations.

Value

A vector with the fitted configuration.

Author(s)

Michael Hahsler with code from Patrick Mair (from smacof).

References

Mair P., De Leeuw J. (2015). Unidimensional scaling. In Wiley StatsRef: Statistics Reference Online, Wiley, New York.

See Also

uniscale in smacof.

Examples

data(SupremeCourt)

d <- as.dist(SupremeCourt)

sc <- uniscale(d)
sc

orderplot(sc)

[Package seriation version 1.2-3 Index]